2002-07-21  Bart Oldeman  <bart.oldeman@btinternet.com>
	* include/shlobj.h (FCIDM_MENU_FAVORITES): Remove bogus
	character.
	* include/winsock2.h (struct sockaddr): Use __int64 instead of
	long long.
	* include/kernel32.c (GetCurrentFiber, GetFiberData): Watcom
	does not need external library functions.

+++ winsup/w32api/include/winsock2.h
@@ -333,7 +333,7 @@ struct sockaddr {
 struct sockaddr_storage {
     short ss_family;
     char __ss_pad1[6];    /* pad to 8 */
-    long long __ss_align; /* force alignment */
+    __int64 __ss_align; /* force alignment */
     char __ss_pad2[112];  /* pad to 128 */
 };
